% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcode.R
\name{dcode}
\alias{dcode}
\title{Code/Decode Commands}
\usage{
dcode(text)
}
\arguments{
\item{dcode}{vector of characters that should be coded/decoded.}
}
\description{
Code/Decode Commands
}
\author{
Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}

