% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommands.R
\name{get_commands}
\alias{get_commands}
\title{Read and Decode Commands}
\usage{
get_commands(fname)
}
\arguments{
\item{fname}{Name of a file to read and transform.}
}
\description{
Read file with commands from a directory and use the 'decode' function to decode it.
}
\author{
Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}

