% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_BB.R
\name{theme_BB}
\alias{theme_BB}
\title{GG Theme for Plots for Project Beta Bit}
\usage{
theme_BB(...)
}
\arguments{
\item{...}{Will be passed to `theme_bw()`}
}
\description{
GG Theme for Plots for Project Beta Bit
}
\author{
Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}

